+++ boxkeeper_v0.11/app/src/main/java/com/example/boxkeeper/data/repository/SearchRepository.java
@@ -2,7 +2,7 @@ package com.example.boxkeeper.data.repository;
 
 import androidx.lifecycle.LiveData;
 
-import com.example.boxkeeper.data.api.TrackingInfo;
+import com.example.boxkeeper.data.model.TrackingInfo;
 
 public interface SearchRepository {
     LiveData<TrackingInfo> getTrackingInfo(String apiKey, String courierCode, String invoiceNumber);
