+++ boxkeeper_v0.11/app/src/main/java/com/example/boxkeeper/data/repository/SearchRepositoryImpl.java
@@ -3,8 +3,8 @@ package com.example.boxkeeper.data.repository;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 
-import com.example.boxkeeper.data.api.SweetTrackerService;
-import com.example.boxkeeper.data.api.TrackingInfo;
+import com.example.boxkeeper.data.remote.SearchRemoteDataSource;
+import com.example.boxkeeper.data.model.TrackingInfo;
 import com.example.boxkeeper.data.api.retrofitClient;
 
 import retrofit2.Call;
@@ -12,7 +12,7 @@ import retrofit2.Callback;
 import retrofit2.Response;
 
 public class SearchRepositoryImpl implements SearchRepository {
-    private final SweetTrackerService sweetTrackerService;
+    private final SearchRemoteDataSource sweetTrackerService;
 
     public SearchRepositoryImpl() {
         sweetTrackerService = retrofitClient.getService();
