+++ boxkeeper_v0.11/app/src/main/java/com/example/boxkeeper/ui/call/CallSharedViewModel.java
@@ -6,6 +6,8 @@ import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
+import com.example.boxkeeper.ui.call.model.CallModel;
+
 import java.util.ArrayList;
 import java.util.List;
