+++ boxkeeper_v0.11/app/src/main/java/com/example/boxkeeper/ui/call/edit/CallInputType.java
@@ -1,4 +1,4 @@
-package com.example.boxkeeper.ui.call;
+package com.example.boxkeeper.ui.call.edit;
 
 public enum CallInputType {
     ADD, EDIT;
