+++ boxkeeper_v0.11/app/src/main/java/com/example/boxkeeper/ui/list/ImageRepository.java
@@ -2,6 +2,8 @@ package com.example.boxkeeper.ui.list;
 
 import androidx.lifecycle.LiveData;
 
+import com.example.boxkeeper.ui.list.model.ImageModel;
+
 import java.util.List;
 
 public interface ImageRepository {
