+++ boxkeeper_v0.11/app/src/main/java/com/example/boxkeeper/ui/list/ImageViewModel.java
@@ -1,9 +1,9 @@
 package com.example.boxkeeper.ui.list;
 
-import android.util.Log;
-
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.ViewModel;
+
+import com.example.boxkeeper.ui.list.model.ImageModel;
 
 import java.util.List;
