+++ boxkeeper_v0.11/app/src/main/java/com/example/boxkeeper/ui/list/model/ImageModel.java
@@ -1,4 +1,4 @@
-package com.example.boxkeeper.ui.list;
+package com.example.boxkeeper.ui.list.model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
